requirement.txt
numpy 
pandas
streamlit
google.generativeai

